This shouldn't have been on the rdac branch, but I'm not good at this Git thing yet.
I added a lot of code to handle operation of the menu via the buttons and also getting an LED light to come on and flash when there are alarms/flashing alarms. The light is steady whenever the alarms aren't flashing. Still need to improve the logic behind when it flashes and also when it should be turned off completely. Yay for buttons though!!